# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Akwaaba  API",
        "description": " description",
        "contact": {
            "email": "dawud@effectstudios.co"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/business-data/create": {
            "post": {
                "tags": [
                    "BusinessData"
                ],
                "summary": "Create Business Data ",
                "description": "Create Business Data",
                "operationId": "createBusinessData",
                "requestBody": {
                    "description": "Create a Business Data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "business_name",
                                    "address",
                                    "experience"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "your email",
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "business_name": {
                                        "description": "your Business Name",
                                        "type": "string",
                                        "example": "trapzium"
                                    },
                                    "address": {
                                        "description": "your business address",
                                        "type": "string",
                                        "example": "Dansoman Roundabout, Accra"
                                    },
                                    "experience": {
                                        "description": "Experience",
                                        "type": "string",
                                        "example": "Food & nightlife, sports events, talks & workshop"
                                    },
                                    "location": {
                                        "properties": {
                                            "street_number": {
                                                "type": "string",
                                                "example": "213"
                                            },
                                            "route": {
                                                "type": "string",
                                                "example": "Hamburgo"
                                            },
                                            "political": {
                                                "type": "string",
                                                "example": "Juárez"
                                            },
                                            "administrative_area_level_1": {
                                                "type": "string",
                                                "example": "Hamburgo"
                                            },
                                            "country": {
                                                "type": "string",
                                                "example": "Mexico"
                                            },
                                            "postal_code": {
                                                "type": "string",
                                                "example": "06600"
                                            },
                                            "lat": {
                                                "type": "string",
                                                "example": "19.4245"
                                            },
                                            "lon": {
                                                "type": "string",
                                                "example": "-99.1688"
                                            },
                                            "address": {
                                                "type": "string",
                                                "example": "Hamburgo 213, Juárez, Mexico City, CDMX, Mexico"
                                            },
                                            "city": {
                                                "type": "string",
                                                "example": "Ciudad de México"
                                            },
                                            "is_active": {
                                                "type": "string",
                                                "example": "true"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Record Added Successfully."
                                        },
                                        "business_data": {
                                            "$ref": "#/components/schemas/BusinessData"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/personal-data/create": {
            "post": {
                "tags": [
                    "PersonalData"
                ],
                "summary": "Create Personal Data ",
                "description": "Create Personal Data",
                "operationId": "createPersonalData",
                "requestBody": {
                    "description": "Create a Personal Data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "first_name",
                                    "last_name",
                                    "experience"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "your email",
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "first_name": {
                                        "description": "your first name",
                                        "type": "string",
                                        "example": "mark"
                                    },
                                    "last_name": {
                                        "description": "your last name",
                                        "type": "string",
                                        "example": "Benzy"
                                    },
                                    "experience": {
                                        "description": "Experience",
                                        "type": "string",
                                        "example": "Food & nightlife, sports events, talks & workshop"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Record Added Successfully."
                                        },
                                        "personal_data": {
                                            "$ref": "#/components/schemas/PersonalData"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "BusinessData": {
                "required": [
                    "email",
                    "business_name",
                    "address",
                    "experience"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "11245"
                    },
                    "email": {
                        "description": "your email",
                        "type": "string",
                        "readOnly": "false",
                        "example": "user@example.com"
                    },
                    "type": {
                        "description": "record type",
                        "type": "string",
                        "readOnly": "false",
                        "example": "business"
                    },
                    "business_name": {
                        "description": "your Business Name",
                        "type": "string",
                        "readOnly": "false",
                        "example": "trapzium"
                    },
                    "address": {
                        "description": "your business address",
                        "type": "string",
                        "readOnly": "false",
                        "example": "Dansoman Roundabout, Accra"
                    },
                    "experience": {
                        "description": "Experience",
                        "type": "string",
                        "example": "Food & nightlife, sports events, talks & workshop"
                    },
                    "location": {
                        "properties": {
                            "street_number": {
                                "type": "string",
                                "example": "213"
                            },
                            "route": {
                                "type": "string",
                                "example": "Hamburgo"
                            },
                            "political": {
                                "type": "string",
                                "example": "Juárez"
                            },
                            "administrative_area_level_1": {
                                "type": "string",
                                "example": "Hamburgo"
                            },
                            "country": {
                                "type": "string",
                                "example": "Mexico"
                            },
                            "postal_code": {
                                "type": "string",
                                "example": "06600"
                            },
                            "lat": {
                                "type": "string",
                                "example": "19.4245"
                            },
                            "lon": {
                                "type": "string",
                                "example": "-99.1688"
                            },
                            "address": {
                                "type": "string",
                                "example": "Hamburgo 213, Juárez, Mexico City, CDMX, Mexico"
                            },
                            "city": {
                                "type": "string",
                                "example": "Ciudad de México"
                            },
                            "is_active": {
                                "type": "string",
                                "example": "true"
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BusinessData"
                }
            },
...
            "PersonalData": {
                "required": [
                    "email",
                    "first_name",
                    "last_name",
                    "experience"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "11245"
                    },
                    "email": {
                        "description": "your email",
                        "type": "string",
                        "readOnly": "false",
                        "example": "user@example.com"
                    },
                    "first_name": {
                        "description": "your first name",
                        "type": "string",
                        "readOnly": "false",
                        "example": "mark"
                    },
                    "last_name": {
                        "description": "your last name",
                        "type": "string",
                        "readOnly": "false",
                        "example": "Benzy"
                    },
                    "experience": {
                        "description": "Experience",
                        "type": "string",
                        "example": "Food & nightlife, sports events, talks & workshop"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PersonalData"
                }
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}